turn sound into a class thats attached to the game state
